+++ directdm/wilson_coefficients.py
@@ -43,7 +43,7 @@ class WC_3flavor(object):
         """ Class for Wilson coefficients in 3 flavor QCD x QED plus DM.
 
         The first argument should be a dictionary for the initial conditions
-        of the 2 + 24 + 4 + 36 + 4 + 48 + 6 + 1 + 12 = 137 
+        of the 2 + 24 + 4 + 36 + 4 + 24 + 6 + 1 + 12 = 113
         dimension-five to dimension-eight three-flavor-QCD Wilson coefficients of the form
         {'C51' : value, 'C52' : value, ...}. 
         An arbitrary number of them can be given; the default values are zero. 
@@ -72,10 +72,6 @@ class WC_3flavor(object):
                              'C716u', 'C716d', 'C716s', 'C716e', 'C716mu', 'C716tau',
                              'C717u', 'C717d', 'C717s', 'C717e', 'C717mu', 'C717tau',
                              'C718u', 'C718d', 'C718s', 'C718e', 'C718mu', 'C718tau',
-                             'C719u', 'C719d', 'C719s', 'C719e', 'C719mu', 'C719tau',
-                             'C720u', 'C720d', 'C720s', 'C720e', 'C720mu', 'C720tau',
-                             'C721u', 'C721d', 'C721s', 'C721e', 'C721mu', 'C721tau',
-                             'C722u', 'C722d', 'C722s', 'C722e', 'C722mu', 'C722tau',
                              'C723u', 'C723d', 'C723s', 'C723e', 'C723mu', 'C723tau',
                              'C725',
                              'C81u', 'C81d', 'C81s', 'C82u', 'C82d', 'C82s'
@@ -89,10 +85,6 @@ class WC_3flavor(object):
                              'C77u', 'C77d', 'C77s', 'C77e', 'C77mu', 'C77tau', 
                              'C78u', 'C78d', 'C78s', 'C78e', 'C78mu', 'C78tau',
                              'C711', 'C712', 'C713', 'C714',
-                             'C715u', 'C715d', 'C715s', 'C715e', 'C715mu', 'C715tau', 
-                             'C716u', 'C716d', 'C716s', 'C716e', 'C716mu', 'C716tau',
-                             'C717u', 'C717d', 'C717s', 'C717e', 'C717mu', 'C717tau', 
-                             'C718u', 'C718d', 'C718s', 'C718e', 'C718mu', 'C718tau',
                              'C82u', 'C82d', 'C82s', 'C84u', 'C84d', 'C84s'
 
         Complex Scalar:      'C61u', 'C61d', 'C61s', 'C61e', 'C61mu', 'C61tau', 
@@ -100,9 +92,9 @@ class WC_3flavor(object):
                              'C63u', 'C63d', 'C63s', 'C63e', 'C63mu', 'C63tau', 
                              'C64u', 'C64d', 'C64s', 'C64e', 'C64mu', 'C64tau',
                              'C65', 'C66', 'C67', 'C68' 
-                             'C81u', 'C81d', 'C81s', 'C82u', 'C82d', 'C82s',
                              'C69u', 'C69d', 'C69s', 'C69e', 'C69mu', 'C69tau', 
                              'C610'
+                             'C81u', 'C81d', 'C81s', 'C82u', 'C82d', 'C82s',
 
         Real Scalar:         'C63u', 'C63d', 'C63s', 'C63e', 'C63mu', 'C63tau', 
                              'C64u', 'C64d', 'C64s', 'C64e', 'C64mu', 'C64tau',
@@ -172,10 +164,6 @@ class WC_3flavor(object):
                                  'C716u', 'C716d', 'C716s', 'C716e', 'C716mu', 'C716tau',
                                  'C717u', 'C717d', 'C717s', 'C717e', 'C717mu', 'C717tau',
                                  'C718u', 'C718d', 'C718s', 'C718e', 'C718mu', 'C718tau',
-                                 'C719u', 'C719d', 'C719s', 'C719e', 'C719mu', 'C719tau',
-                                 'C720u', 'C720d', 'C720s', 'C720e', 'C720mu', 'C720tau', 
-                                 'C721u', 'C721d', 'C721s', 'C721e', 'C721mu', 'C721tau',
-                                 'C722u', 'C722d', 'C722s', 'C722e', 'C722mu', 'C722tau',
                                  'C723u', 'C723d', 'C723s', 'C723e', 'C723mu', 'C723tau',
                                  'C725']
 
@@ -191,17 +179,13 @@ class WC_3flavor(object):
                                  'C77u', 'C77d', 'C77s', 'C77e', 'C77mu', 'C77tau',
                                  'C78u', 'C78d', 'C78s', 'C78e', 'C78mu', 'C78tau',
                                  'C711', 'C712', 'C713', 'C714',
-                                 'C715u', 'C715d', 'C715s', 'C715e', 'C715mu', 'C715tau',
-                                 'C716u', 'C716d', 'C716s', 'C716e', 'C716mu', 'C716tau',
-                                 'C717u', 'C717d', 'C717s', 'C717e', 'C717mu', 'C717tau',
-                                 'C718u', 'C718d', 'C718s', 'C718e', 'C718mu', 'C718tau',
                                  'C723u', 'C723d', 'C723s', 'C723e', 'C723mu', 'C723tau',
                                  'C725']
 
             self.wc8_name_list = ['C82u', 'C82d', 'C82s', 'C84u', 'C84d', 'C84s']
 
             # The list of indices to be deleted from the QCD/QED ADM because of less operators
-            del_ind_list = np.r_[np.s_[0:8], np.s_[14:20], np.s_[54:66], np.s_[94:118]]
+            del_ind_list = np.r_[np.s_[0:8], np.s_[14:20], np.s_[54:66], np.s_[70:94]]
             # The list of indices to be deleted from the dim.8 ADM because of less operators
             del_ind_list_dim_8 = np.r_[np.s_[0:3], np.s_[6:9]]
 
@@ -219,7 +203,7 @@ class WC_3flavor(object):
 
             # The list of indices to be deleted from the QCD/QED ADM because of less operators
             del_ind_list = np.r_[np.s_[0:2], np.s_[8:14], np.s_[20:26], np.s_[27:28], np.s_[29:30],\
-                                 np.s_[36:42], np.s_[48:66], np.s_[67:68], np.s_[69:70], np.s_[70:118]]
+                                 np.s_[36:42], np.s_[48:66], np.s_[67:68], np.s_[69:70], np.s_[70:94]]
             # The list of indices to be deleted from the dim.8 ADM because of less operators
             del_ind_list_dim_8 = np.r_[np.s_[0:3], np.s_[6:9]]
 
@@ -235,7 +219,7 @@ class WC_3flavor(object):
 
             # The list of indices to be deleted from the QCD/QED ADM because of less operators
             del_ind_list = np.r_[np.s_[0:26], np.s_[27:28], np.s_[29:30], np.s_[36:42],\
-                                 np.s_[48:66], np.s_[67:68], np.s_[69:70], np.s_[70:118]]
+                                 np.s_[48:66], np.s_[67:68], np.s_[69:70], np.s_[70:94]]
 
         self.coeff_dict = {}
 
@@ -548,7 +532,6 @@ class WC_3flavor(object):
                       + FGp*c3mu_dict['C71']\
                       + FSup*c3mu_dict['C75u'] + FSdp*c3mu_dict['C75d'] + FSsp*c3mu_dict['C75s']\
                       - alpha/(2*np.pi*DM_mass)*c3mu_dict['C51']\
-                      + 2*DM_mass * (F1up*c3mu_dict['C715u'] + F1dp*c3mu_dict['C715d'] + F1sp*c3mu_dict['C715s'])\
                       + FTW2up*c3mu_dict['C723u']\
                       + FTW2dp*c3mu_dict['C723d']\
                       + FTW2sp*c3mu_dict['C723s']\
@@ -560,17 +543,18 @@ class WC_3flavor(object):
                            + FAsp*(c3mu_dict['C64s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C84s']))\
                       - 2*alpha/np.pi * self.ip['mup']/mN * c3mu_dict['C51']\
                       + 8*(FT0up*c3mu_dict['C79u'] + FT0dp*c3mu_dict['C79d'] + FT0sp*c3mu_dict['C79s']),
-            'cNR5p' : - 2*mN * (F1up*c3mu_dict['C719u'] + F1dp*c3mu_dict['C719d'] + F1sp*c3mu_dict['C719s']),
+            'cNR5p' : - 2*mN * (  F1up*c3mu_dict['C715u']\
+                                + F1dp*c3mu_dict['C715d']\
+                                + F1sp*c3mu_dict['C715s']),
             'cNR6p' : mN/DM_mass * FGtildep * c3mu_dict['C74']\
-                      -2*mN*((F1up+F2up)*c3mu_dict['C719u']\
-                             + (F1dp+F2dp)*c3mu_dict['C719d']\
-                             + (F1sp+F2sp)*c3mu_dict['C719s'])\
+                      -2*mN*(  (F1up+F2up)*c3mu_dict['C715u']\
+                             + (F1dp+F2dp)*c3mu_dict['C715d']\
+                             + (F1sp+F2sp)*c3mu_dict['C715s'])\
                       + mN/DM_mass*F2sp*(c3mu_dict['C61s']
                                          - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C81s']),
             'cNR7p' : - 2*(  FAup*(c3mu_dict['C63u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C83u'])\
                            + FAdp*(c3mu_dict['C63d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C83d'])\
-                           + FAsp*(c3mu_dict['C63s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C83s']))\
-                      - 4*DM_mass * (FAup*c3mu_dict['C717u'] + FAdp*c3mu_dict['C717d'] + FAsp*c3mu_dict['C717s']),
+                           + FAsp*(c3mu_dict['C63s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C83s'])),
             'cNR8p' : 2*(  F1up*(c3mu_dict['C62u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C82u'])\
                          + F1dp*(c3mu_dict['C62d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C82d'])),
             'cNR9p' : 2*(  (F1up+F2up)*(c3mu_dict['C62u']\
@@ -583,7 +567,9 @@ class WC_3flavor(object):
                               + FAdp*(c3mu_dict['C63d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C83d'])\
                               + FAsp*(c3mu_dict['C63s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C83s']))\
                              /DM_mass\
-                      - 4*mN * (FAup*c3mu_dict['C721u'] + FAdp*c3mu_dict['C721d'] + FAsp*c3mu_dict['C721s']),
+                      - 4*mN * (  FAup*c3mu_dict['C717u']\
+                                + FAdp*c3mu_dict['C717d']\
+                                + FAsp*c3mu_dict['C717s']),
             'cNR10p' : FGtildep * c3mu_dict['C73']\
                        -2*mN/DM_mass * (FT0up*c3mu_dict['C710u']\
                                         + FT0dp*c3mu_dict['C710d']\
@@ -595,14 +581,14 @@ class WC_3flavor(object):
                         + 2*((FT0up-FT1up)*c3mu_dict['C710u']\
                              + (FT0dp-FT1dp)*c3mu_dict['C710d']\
                              + (FT0sp-FT1sp)*c3mu_dict['C710s'])\
-                        - 2*mN * (  F1up*(c3mu_dict['C716u']+c3mu_dict['C720u'])\
-                                  + F1dp*(c3mu_dict['C716d']+c3mu_dict['C720d'])\
-                                  + F1sp*(c3mu_dict['C716s']+c3mu_dict['C720s'])),
+                        - 2*mN * (  F1up*(c3mu_dict['C716u'])\
+                                  + F1dp*(c3mu_dict['C716d'])\
+                                  + F1sp*(c3mu_dict['C716s'])),
             'cNR12p' : -8*(FT0up*c3mu_dict['C710u'] + FT0dp*c3mu_dict['C710d'] + FT0sp*c3mu_dict['C710s']),
             'cNR13p' : 0.,
-            'cNR14p' : + 4*mN * (  FAup*(c3mu_dict['C718u']+c3mu_dict['C722u'])\
-                                 + FAdp*(c3mu_dict['C718d']+c3mu_dict['C722d'])\
-                                 + FAsp*(c3mu_dict['C718s']+c3mu_dict['C722s'])),
+            'cNR14p' : + 4*mN * (  FAup*(c3mu_dict['C718u'])\
+                                 + FAdp*(c3mu_dict['C718d'])\
+                                 + FAsp*(c3mu_dict['C718s'])),
 
             'cNR6pip' : mN/DM_mass * (FPup_pion*c3mu_dict['C78u'] + FPdp_pion*c3mu_dict['C78d'])\
                        + FPpup_pion*(c3mu_dict['C64u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C84u'])\
@@ -625,14 +611,14 @@ class WC_3flavor(object):
             'cNR6bq2p' : -mN**2* (- 2*alpha/np.pi * self.ip['mup']/mN * c3mu_dict['C51']),
             'cNR11bq2p' : mN* (2*alpha/np.pi*c3mu_dict['C52']),
 
-            'cNR1q2p' : (  F1up*c3mu_dict['C719u']\
-                         + F1dp*c3mu_dict['C719d']\
-                         + F1sp*c3mu_dict['C719s'])/(2*DM_mass)\
+            'cNR1q2p' : (  F1up*c3mu_dict['C715u']\
+                         + F1dp*c3mu_dict['C715d']\
+                         + F1sp*c3mu_dict['C715s'])/(2*DM_mass)\
                         + (F1spslope - F2sp / mN**2/4)
                           * (c3mu_dict['C61s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C81s']),
-            'cNR4q2p' : 2*((F1up+F2up)*c3mu_dict['C719u']\
-                           + (F1dp+F2dp)*c3mu_dict['C719d']\
-                           + (F1sp+F2sp)*c3mu_dict['C719s'])/mN\
+            'cNR4q2p' : 2*(  (F1up+F2up)*c3mu_dict['C715u']\
+                           + (F1dp+F2dp)*c3mu_dict['C715d']\
+                           + (F1sp+F2sp)*c3mu_dict['C715s'])/mN\
                         - 1/mN/DM_mass * F2sp
                           * (c3mu_dict['C61s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C81s']),
 
@@ -643,7 +629,6 @@ class WC_3flavor(object):
                       + F1dn*(c3mu_dict['C61d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C81d'])\
                       + FGn*c3mu_dict['C71']\
                       + FSun*c3mu_dict['C75u'] + FSdn*c3mu_dict['C75d'] + FSsn*c3mu_dict['C75s']\
-                      + 2*DM_mass * (F1un*c3mu_dict['C715u'] + F1dn*c3mu_dict['C715d'] + F1sn*c3mu_dict['C715s'])\
                       + FTW2un*c3mu_dict['C723u']\
                       + FTW2dn*c3mu_dict['C723d']\
                       + FTW2sn*c3mu_dict['C723s']\
@@ -655,17 +640,18 @@ class WC_3flavor(object):
                            + FAsn*(c3mu_dict['C64s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C84s']))\
                       - 2*alpha/np.pi * self.ip['mun']/mN * c3mu_dict['C51']\
                       + 8*(FT0un*c3mu_dict['C79u'] + FT0dn*c3mu_dict['C79d'] + FT0sn*c3mu_dict['C79s']),
-            'cNR5n' : - 2*mN * (F1un*c3mu_dict['C719u'] + F1dn*c3mu_dict['C719d'] + F1sn*c3mu_dict['C719s']),
+            'cNR5n' : - 2*mN * (  F1un*c3mu_dict['C715u']\
+                                + F1dn*c3mu_dict['C715d']\
+                                + F1sn*c3mu_dict['C715s']),
             'cNR6n' : mN/DM_mass * FGtilden * c3mu_dict['C74']\
-                      -2*mN*((F1un+F2un)*c3mu_dict['C719u']\
-                             + (F1dn+F2dn)*c3mu_dict['C719d']\
-                             + (F1sn+F2sn)*c3mu_dict['C719s'])\
+                      -2*mN*(  (F1un+F2un)*c3mu_dict['C715u']\
+                             + (F1dn+F2dn)*c3mu_dict['C715d']\
+                             + (F1sn+F2sn)*c3mu_dict['C715s'])\
                       + mN/DM_mass * F2sn
                         * (c3mu_dict['C61s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C81s']),
             'cNR7n' : - 2*(  FAun*(c3mu_dict['C63u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C83u'])\
                            + FAdn*(c3mu_dict['C63d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C83d'])\
-                           + FAsn*(c3mu_dict['C63s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C83s']))\
-                      - 4*DM_mass * (FAun*c3mu_dict['C717u'] + FAdn*c3mu_dict['C717d']+ FAsn*c3mu_dict['C717s']),
+                           + FAsn*(c3mu_dict['C63s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C83s'])),
             'cNR8n' : 2*(  F1un*(c3mu_dict['C62u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C82u'])\
                          + F1dn*(c3mu_dict['C62d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C82d'])),
             'cNR9n' : 2*(  (F1un+F2un)*(c3mu_dict['C62u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C82u'])\
@@ -675,9 +661,9 @@ class WC_3flavor(object):
                               + FAdn*(c3mu_dict['C63d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C83d'])\
                               + FAsn*(c3mu_dict['C63s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C83s']))\
                              /DM_mass\
-                      - 4*mN * (FAun*c3mu_dict['C721u']\
-                                + FAdn*c3mu_dict['C721d']\
-                                + FAsn*c3mu_dict['C721s']),
+                      - 4*mN * (  FAun*c3mu_dict['C717u']\
+                                + FAdn*c3mu_dict['C717d']\
+                                + FAsn*c3mu_dict['C717s']),
             'cNR10n' : FGtilden * c3mu_dict['C73']\
                      -2*mN/DM_mass * (FT0un*c3mu_dict['C710u']\
                                       + FT0dn*c3mu_dict['C710d']\
@@ -689,14 +675,14 @@ class WC_3flavor(object):
                        + 2*((FT0un-FT1un)*c3mu_dict['C710u']\
                             + (FT0dn-FT1dn)*c3mu_dict['C710d']\
                             + (FT0sn-FT1sn)*c3mu_dict['C710s'])\
-                       - 2*mN * (  F1un*(c3mu_dict['C716u']+c3mu_dict['C720u'])\
-                                 + F1dn*(c3mu_dict['C716d']+c3mu_dict['C720d'])\
-                                 + F1sn*(c3mu_dict['C716s']+c3mu_dict['C720s'])),
+                       - 2*mN * (  F1un*(c3mu_dict['C716u'])\
+                                 + F1dn*(c3mu_dict['C716d'])\
+                                 + F1sn*(c3mu_dict['C716s'])),
             'cNR12n' : -8*(FT0un*c3mu_dict['C710u'] + FT0dn*c3mu_dict['C710d'] + FT0sn*c3mu_dict['C710s']),
             'cNR13n' : 0.,
-            'cNR14n' :           + 4*mN * (  FAun*(c3mu_dict['C718u']+c3mu_dict['C722u'])\
-                                 + FAdn*(c3mu_dict['C718d']+c3mu_dict['C722d'])\
-                                 + FAsn*(c3mu_dict['C718s']+c3mu_dict['C722s'])),
+            'cNR14n' : + 4*mN * (  FAun*(c3mu_dict['C718u'])\
+                                 + FAdn*(c3mu_dict['C718d'])\
+                                 + FAsn*(c3mu_dict['C718s'])),
     
             'cNR6pin' : mN/DM_mass * (FPun_pion*c3mu_dict['C78u'] + FPdn_pion*c3mu_dict['C78d'])\
                        + FPpun_pion*(c3mu_dict['C64u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C84u'])\
@@ -719,38 +705,33 @@ class WC_3flavor(object):
             'cNR6bq2n' : -mN**2 * (- 2*alpha/np.pi * self.ip['mun']/mN * c3mu_dict['C51']),
             'cNR11bq2n' : 0,
 
-            'cNR1q2n' : (  F1un*c3mu_dict['C719u']\
-                         + F1dn*c3mu_dict['C719d']\
-                         + F1sn*c3mu_dict['C719s'])/(2*DM_mass)\
+            'cNR1q2n' : (  F1un*c3mu_dict['C715u']\
+                         + F1dn*c3mu_dict['C715d']\
+                         + F1sn*c3mu_dict['C715s'])/(2*DM_mass)\
                         + (F1snslope - F2sn / mN**2/4)
                           * (c3mu_dict['C61s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C81s']),
-            'cNR4q2n' : 2*((F1un+F2un)*c3mu_dict['C719u']\
-                           + (F1dn+F2dn)*c3mu_dict['C719d']\
-                           + (F1sn+F2sn)*c3mu_dict['C719s'])/mN\
+            'cNR4q2n' : 2*(  (F1un+F2un)*c3mu_dict['C715u']\
+                           + (F1dn+F2dn)*c3mu_dict['C715d']\
+                           + (F1sn+F2sn)*c3mu_dict['C715s'])/mN\
                         - 1/mN/DM_mass * F2sn
                           * (c3mu_dict['C61s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C81s'])
             }
 
             if NLO:
                 my_cNR_dict['cNR5p'] = my_cNR_dict['cNR5p']\
-                                       - 2*mN * (F1un*c3mu_dict['C719u']\
-                                                 + F1dn*c3mu_dict['C719d']\
-                                                 + F1sn*c3mu_dict['C719s'])\
-                                       + 2*((FT0up-FT1up)*c3mu_dict['C79u']\
+                                       + 2*(  (FT0up-FT1up)*c3mu_dict['C79u']\
                                             + (FT0dp-FT1dp)*c3mu_dict['C79d']\
                                             + (FT0sp-FT1sp)*c3mu_dict['C79s'])
                 my_cNR_dict['cNR1q2p'] = my_cNR_dict['cNR1q2p']\
-                                         - ((FT0up-FT1up)*c3mu_dict['C79u']\
+                                         - (  (FT0up-FT1up)*c3mu_dict['C79u']\
                                             + (FT0dp-FT1dp)*c3mu_dict['C79d']\
                                             + (FT0sp-FT1sp)*c3mu_dict['C79s'])/(2*DM_mass*mN)
                 my_cNR_dict['cNR5n'] = my_cNR_dict['cNR5n']\
-                                       - 2*mN * (F1un*c3mu_dict['C719u']\
-                                                 + F1dn*c3mu_dict['C719d'] + F1sn*c3mu_dict['C719s'])\
-                                       + 2*((FT0un-FT1un)*c3mu_dict['C79u']\
+                                       + 2*(  (FT0un-FT1un)*c3mu_dict['C79u']\
                                             + (FT0dn-FT1dn)*c3mu_dict['C79d']\
                                             + (FT0sn-FT1sn)*c3mu_dict['C79s'])
                 my_cNR_dict['cNR1q2n'] = my_cNR_dict['cNR1q2n']\
-                                         - ((FT0un-FT1un)*c3mu_dict['C79u']\
+                                         - (  (FT0un-FT1un)*c3mu_dict['C79u']\
                                             + (FT0dn-FT1dn)*c3mu_dict['C79d']\
                                             + (FT0sn-FT1sn)*c3mu_dict['C79s'])/(2*DM_mass*mN)
 
@@ -758,16 +739,15 @@ class WC_3flavor(object):
         if self.DM_type == "M":
             my_cNR_dict = {
             'cNR1p' : FGp*c3mu_dict['C71']\
-                      + FSup*c3mu_dict['C75u'] + FSdp*c3mu_dict['C75d'] + FSsp*c3mu_dict['C75s']\
-                      + 2*DM_mass * (F1up*c3mu_dict['C715u'] + F1dp*c3mu_dict['C715d'] + F1sp*c3mu_dict['C715s']),
-            'cNR2p' : 0,
-            'cNR3p' : 0,
+                      + FSup*c3mu_dict['C75u'] + FSdp*c3mu_dict['C75d'] + FSsp*c3mu_dict['C75s'],
+            'cNR2p' : 0.,
+            'cNR3p' : 0.,
             'cNR4p' : - 4*(  FAup*(c3mu_dict['C64u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C84u'])\
                            + FAdp*(c3mu_dict['C64d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C84d'])\
                            + FAsp*(c3mu_dict['C64s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C84s'])),
             'cNR5p' : 0,
             'cNR6p' : mN/DM_mass * FGtildep * c3mu_dict['C74'],
-            'cNR7p' : - 4*DM_mass * (FAup*c3mu_dict['C717u'] + FAdp*c3mu_dict['C717d'] + FAsp*c3mu_dict['C717s']),
+            'cNR7p' : 0,
             'cNR8p' : 2*(  F1up*(c3mu_dict['C62u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C82u'])\
                          + F1dp*(c3mu_dict['C62d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C82d'])),
             'cNR9p' : 2*(  (F1up+F2up)*(c3mu_dict['C62u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C82u'])\
@@ -777,15 +757,10 @@ class WC_3flavor(object):
             'cNR11p' : - mN/DM_mass * (FSup*c3mu_dict['C76u']\
                                        + FSdp*c3mu_dict['C76d']\
                                        + FSsp*c3mu_dict['C76s'])\
-                       - mN/DM_mass * FGp * c3mu_dict['C72']\
-                       - 2*mN * (  F1up*c3mu_dict['C716u']\
-                                   + F1dp*c3mu_dict['C716d']\
-                                   + F1sp*c3mu_dict['C716s']),
-            'cNR12p' : 0,
+                       - mN/DM_mass * FGp * c3mu_dict['C72'],
+            'cNR12p' : 0.,
             'cNR13p' : 0.,
-            'cNR14p' : + 4*mN * (FAup*c3mu_dict['C718u']\
-                                 + FAdp*c3mu_dict['C718d']\
-                                 + FAsp*c3mu_dict['C718s']),
+            'cNR14p' : 0.,
     
             'cNR6pip' : mN/DM_mass * (FPup_pion*c3mu_dict['C78u'] + FPdp_pion*c3mu_dict['C78d'])\
                        + FPpup_pion*(c3mu_dict['C64u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C84u'])\
@@ -804,27 +779,26 @@ class WC_3flavor(object):
             'cNR10q2pip' : FGtildep_pion * c3mu_dict['C73'],
             'cNR10q2etap' : FGtildep_eta * c3mu_dict['C73'],
     
-            'cNR5bq2p' : 0,
-            'cNR6bq2p' : 0,
-            'cNR11bq2p' : 0,
+            'cNR5bq2p' : 0.,
+            'cNR6bq2p' : 0.,
+            'cNR11bq2p' : 0.,
 
-            'cNR1q2p' : 0,
-            'cNR4q2p' : 0,
+            'cNR1q2p' : 0.,
+            'cNR4q2p' : 0.,
 
 
 
 
             'cNR1n' :   FGn*c3mu_dict['C71']\
-                      + FSun*c3mu_dict['C75u'] + FSdn*c3mu_dict['C75d'] + FSsn*c3mu_dict['C75s']\
-                      + 2*DM_mass * (F1un*c3mu_dict['C715u'] + F1dn*c3mu_dict['C715d'] + F1sn*c3mu_dict['C715s']),
-            'cNR2n' : 0,
-            'cNR3n' : 0,
+                      + FSun*c3mu_dict['C75u'] + FSdn*c3mu_dict['C75d'] + FSsn*c3mu_dict['C75s'],
+            'cNR2n' : 0.,
+            'cNR3n' : 0.,
             'cNR4n' : - 4*(  FAun*(c3mu_dict['C64u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C84u'])\
                            + FAdn*(c3mu_dict['C64d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C84d'])\
                            + FAsn*(c3mu_dict['C64s'] - np.sqrt(2)*GF*ms**2 / gs2_2GeV * c3mu_dict['C84s'])),
-            'cNR5n' : 0,
+            'cNR5n' : 0.,
             'cNR6n' : mN/DM_mass * FGtilden * c3mu_dict['C74'],
-            'cNR7n' : - 4*DM_mass * (FAun*c3mu_dict['C717u'] + FAdn*c3mu_dict['C717d'] + FAsn*c3mu_dict['C717s']),
+            'cNR7n' : 0.,
             'cNR8n' : 2*(  F1un*(c3mu_dict['C62u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C82u'])\
                          + F1dn*(c3mu_dict['C62d'] - np.sqrt(2)*GF*md**2 / gs2_2GeV * c3mu_dict['C82d'])),
             'cNR9n' : 2*(  (F1un+F2un)*(c3mu_dict['C62u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C82u'])\
@@ -834,15 +808,10 @@ class WC_3flavor(object):
             'cNR11n' : - mN/DM_mass * (FSun*c3mu_dict['C76u']\
                                        + FSdn*c3mu_dict['C76d']\
                                        + FSsn*c3mu_dict['C76s'])\
-                       - mN/DM_mass * FGn * c3mu_dict['C72']\
-                       - 2*mN * (  F1un*c3mu_dict['C716u']\
-                                   + F1dn*c3mu_dict['C716d']\
-                                   + F1sn*c3mu_dict['C716s']),
-            'cNR12n' : 0,
-            'cNR13p' : 0.,
-            'cNR14n' : + 4*mN * (FAun*c3mu_dict['C718u']\
-                                 + FAdn*c3mu_dict['C718d']\
-                                 + FAsn*c3mu_dict['C718s']),
+                       - mN/DM_mass * FGn * c3mu_dict['C72'],
+            'cNR12n' : 0.,
+            'cNR13n' : 0.,
+            'cNR14n' : 0.,
     
             'cNR6pin' : mN/DM_mass * (FPun_pion*c3mu_dict['C78u'] + FPdn_pion*c3mu_dict['C78d'])\
                        + FPpun_pion*(c3mu_dict['C64u'] - np.sqrt(2)*GF*mu**2 / gs2_2GeV * c3mu_dict['C84u'])\
@@ -861,12 +830,12 @@ class WC_3flavor(object):
             'cNR10q2pin' : FGtilden_pion * c3mu_dict['C73'],
             'cNR10q2etan' : FGtilden_eta * c3mu_dict['C73'],
     
-            'cNR5bq2n' : 0,
-            'cNR6bq2n' : 0,
-            'cNR11bq2n' : 0,
+            'cNR5bq2n' : 0.,
+            'cNR6bq2n' : 0.,
+            'cNR11bq2n' : 0.,
 
-            'cNR1q2n' : 0,
-            'cNR4q2n' : 0
+            'cNR1q2n' : 0.,
+            'cNR4q2n' : 0.
             }
 
 
@@ -1195,7 +1164,7 @@ class WC_4flavor(object):
         """ Class for Wilson coefficients in 4 flavor QCD x QED plus DM.
 
         The argument should be a dictionary for the initial conditions
-        of the 2 + 28 + 4 + 42 + 4 + 56 + 7 + 1 + 6 = 150 dimension-five to dimension-eight
+        of the 2 + 28 + 4 + 42 + 4 + 28 + 7 + 1 + 6 = 122 dimension-five to dimension-eight
         four-flavor-QCD Wilson coefficients (for Dirac DM) of the form
         {'C51' : value, 'C52' : value, ...}. For other DM types there are less coefficients.
         An arbitrary number of them can be given; the default values are zero. 
@@ -1224,10 +1193,6 @@ class WC_4flavor(object):
                              'C716u', 'C716d', 'C716s', 'C716c', 'C716e', 'C716mu', 'C716tau',
                              'C717u', 'C717d', 'C717s', 'C717c', 'C717e', 'C717mu', 'C717tau',
                              'C718u', 'C718d', 'C718s', 'C718c', 'C718e', 'C718mu', 'C718tau',
-                             'C719u', 'C719d', 'C719s', 'C719c', 'C719e', 'C719mu', 'C719tau',
-                             'C720u', 'C720d', 'C720s', 'C720c', 'C720e', 'C720mu', 'C720tau',
-                             'C721u', 'C721d', 'C721s', 'C721c', 'C721e', 'C721mu', 'C721tau',
-                             'C722u', 'C722d', 'C722s', 'C722c', 'C722e', 'C722mu', 'C722tau',
                              'C723u', 'C723d', 'C723s', 'C723c', 'C723e', 'C723mu', 'C723tau',
                              'C725',
                              'C83u', 'C83d', 'C83s', 'C84u', 'C84d', 'C84s'
@@ -1241,10 +1206,6 @@ class WC_4flavor(object):
                              'C77u', 'C77d', 'C77s', 'C77c', 'C77e', 'C77mu', 'C77tau', 
                              'C78u', 'C78d', 'C78s', 'C78c', 'C78e', 'C78mu', 'C78tau',
                              'C711', 'C712', 'C713', 'C714',
-                             'C715u', 'C715d', 'C715s', 'C715c', 'C715e', 'C715mu', 'C715tau', 
-                             'C716u', 'C716d', 'C716s', 'C716c', 'C716e', 'C716mu', 'C716tau',
-                             'C717u', 'C717d', 'C717s', 'C717c', 'C717e', 'C717mu', 'C717tau', 
-                             'C718u', 'C718d', 'C718s', 'C718c', 'C718e', 'C718mu', 'C718tau',
                              'C723u', 'C723d', 'C723s', 'C723c', 'C723e', 'C723mu', 'C723tau',
                              'C725',
 
@@ -1348,10 +1309,6 @@ class WC_4flavor(object):
                                  'C716u', 'C716d', 'C716s', 'C716c', 'C716e', 'C716mu', 'C716tau',
                                  'C717u', 'C717d', 'C717s', 'C717c', 'C717e', 'C717mu', 'C717tau', 
                                  'C718u', 'C718d', 'C718s', 'C718c', 'C718e', 'C718mu', 'C718tau',
-                                 'C719u', 'C719d', 'C719s', 'C719c', 'C719e', 'C719mu', 'C719tau', 
-                                 'C720u', 'C720d', 'C720s', 'C720c', 'C720e', 'C720mu', 'C720tau', 
-                                 'C721u', 'C721d', 'C721s', 'C721c', 'C721e', 'C721mu', 'C721tau', 
-                                 'C722u', 'C722d', 'C722s', 'C722c', 'C722e', 'C722mu', 'C722tau',
                                  'C723u', 'C723d', 'C723s', 'C723c', 'C723e', 'C723mu', 'C723tau',
                                  'C725']
 
@@ -1375,10 +1332,6 @@ class WC_4flavor(object):
                                     'C716u', 'C716d', 'C716s', 'C716e', 'C716mu', 'C716tau',
                                     'C717u', 'C717d', 'C717s', 'C717e', 'C717mu', 'C717tau', 
                                     'C718u', 'C718d', 'C718s', 'C718e', 'C718mu', 'C718tau',
-                                    'C719u', 'C719d', 'C719s', 'C719e', 'C719mu', 'C719tau', 
-                                    'C720u', 'C720d', 'C720s', 'C720e', 'C720mu', 'C720tau', 
-                                    'C721u', 'C721d', 'C721s', 'C721e', 'C721mu', 'C721tau', 
-                                    'C722u', 'C722d', 'C722s', 'C722e', 'C722mu', 'C722tau',
                                     'C723u', 'C723d', 'C723s', 'C723e', 'C723mu', 'C723tau',
                                     'C725']
 
@@ -1391,17 +1344,13 @@ class WC_4flavor(object):
                                  'C77u', 'C77d', 'C77s', 'C77c', 'C77e', 'C77mu', 'C77tau', 
                                  'C78u', 'C78d', 'C78s', 'C78c', 'C78e', 'C78mu', 'C78tau',
                                  'C711', 'C712', 'C713', 'C714',
-                                 'C715u', 'C715d', 'C715s', 'C715c', 'C715e', 'C715mu', 'C715tau', 
-                                 'C716u', 'C716d', 'C716s', 'C716c', 'C716e', 'C716mu', 'C716tau',
-                                 'C717u', 'C717d', 'C717s', 'C717c', 'C717e', 'C717mu', 'C717tau', 
-                                 'C718u', 'C718d', 'C718s', 'C718c', 'C718e', 'C718mu', 'C718tau',
                                  'C723u', 'C723d', 'C723s', 'C723c', 'C723e', 'C723mu', 'C723tau',
                                  'C725']
 
             self.wc8_name_list = ['C82u', 'C82d', 'C82s', 'C84u', 'C84d', 'C84s']
 
             # The list of indices to be deleted from the QCD/QED ADM because of less operators
-            del_ind_list = np.r_[np.s_[0:9], np.s_[16:23], np.s_[62:76], np.s_[108:136]]
+            del_ind_list = np.r_[np.s_[0:9], np.s_[16:23], np.s_[62:76], np.s_[80:108]]
             # The list of indices to be deleted from the dim.8 ADM because of less operators
             del_ind_list_dim_8 = np.r_[np.s_[0:3], np.s_[6:9]]
             # The list of indices to be deleted from the ADT because of less operators (dim.6 part)
@@ -1416,10 +1365,6 @@ class WC_4flavor(object):
                                     'C77u', 'C77d', 'C77s', 'C77e', 'C77mu', 'C77tau',
                                     'C78u', 'C78d', 'C78s', 'C78e', 'C78mu', 'C78tau',
                                     'C711', 'C712', 'C713', 'C714',
-                                    'C715u', 'C715d', 'C715s', 'C715e', 'C715mu', 'C715tau', 
-                                    'C716u', 'C716d', 'C716s', 'C716e', 'C716mu', 'C716tau',
-                                    'C717u', 'C717d', 'C717s', 'C717e', 'C717mu', 'C717tau', 
-                                    'C718u', 'C718d', 'C718s', 'C718e', 'C718mu', 'C718tau',
                                     'C723u', 'C723d', 'C723s', 'C723e', 'C723mu', 'C723tau',
                                     'C725']
 
@@ -1438,7 +1383,7 @@ class WC_4flavor(object):
             # The list of indices to be deleted from the QCD/QED ADM because of less operators
             del_ind_list = [0,1] + [i for i in range(9,16)] + [i for i in range(23,30)]\
                            + [31] + [33] + [i for i in range(41,48)]\
-                           + [i for i in range(55,76)] + [77] + [79] + [i for i in range(80,136)]
+                           + [i for i in range(55,76)] + [77] + [79] + [i for i in range(80,108)]
             # The list of indices to be deleted from the dim.8 ADM because of less operators
             del_ind_list_dim_8 = np.r_[np.s_[0:3], np.s_[6:9]]
             # The list of indices to be deleted from the ADT because of less operators (dim.6 part)
@@ -1467,7 +1412,7 @@ class WC_4flavor(object):
             # The list of indices to be deleted from the QCD/QED ADM because of less operators
             del_ind_list = [i for i in range(0,30)] + [31] + [33] + [i for i in range(41,48)]\
                            + [i for i in range(55,76)]\
-                           + [77] + [79] + [i for i in range(80,136)]
+                           + [77] + [79] + [i for i in range(80,108)]
 
             # The 3-flavor list for matching only
             self.wc_name_list_3f = ['C65', 'C66',
@@ -2150,7 +2095,7 @@ class WC_5flavor(object):
         """ Class for Wilson coefficients in 5 flavor QCD x QED plus DM.
 
         The argument should be a dictionary for the initial conditions of the 
-        2 + 32 + 4 + 48 + 4 + 64 + 8 + 1 + 12 = 175 
+        2 + 32 + 4 + 48 + 4 + 32 + 8 + 1 + 12 = 143 
         dimension-five to dimension-eight five-flavor-QCD Wilson coefficients (for Dirac DM) of the form
         {'C51' : value, 'C52' : value, ...}. For other DM types there are less coefficients.
         A subset of twist-two operators is currently only included for Dirac DM.
@@ -2179,10 +2124,6 @@ class WC_5flavor(object):
                              'C716u', 'C716d', 'C716s', 'C716c', 'C716b', 'C716e', 'C716mu', 'C716tau',
                              'C717u', 'C717d', 'C717s', 'C717c', 'C717b', 'C717e', 'C717mu', 'C717tau', 
                              'C718u', 'C718d', 'C718s', 'C718c', 'C718b', 'C718e', 'C718mu', 'C718tau',
-                             'C719u', 'C719d', 'C719s', 'C719c', 'C719b', 'C719e', 'C719mu', 'C719tau', 
-                             'C720u', 'C720d', 'C720s', 'C720c', 'C720b', 'C720e', 'C720mu', 'C720tau', 
-                             'C721u', 'C721d', 'C721s', 'C721c', 'C721b', 'C721e', 'C721mu', 'C721tau', 
-                             'C722u', 'C722d', 'C722s', 'C722c', 'C722b', 'C722e', 'C722mu', 'C722tau',
                              'C723u', 'C723d', 'C723s', 'C723c', 'C723b', 'C723e', 'C723mu', 'C723tau',
                              'C725',
                              'C81u', 'C81d', 'C81s', 'C82u', 'C82d', 'C82s'
@@ -2196,10 +2137,6 @@ class WC_5flavor(object):
                              'C77u', 'C77d', 'C77s', 'C77c', 'C77b', 'C77e', 'C77mu', 'C77tau', 
                              'C78u', 'C78d', 'C78s', 'C78c', 'C78b', 'C78e', 'C78mu', 'C78tau',
                              'C711', 'C712', 'C713', 'C714',
-                             'C715u', 'C715d', 'C715s', 'C715c', 'C715b', 'C715e', 'C715mu', 'C715tau', 
-                             'C716u', 'C716d', 'C716s', 'C716c', 'C716b', 'C716e', 'C716mu', 'C716tau',
-                             'C717u', 'C717d', 'C717s', 'C717c', 'C717b', 'C717e', 'C717mu', 'C717tau', 
-                             'C718u', 'C718d', 'C718s', 'C718c', 'C718b', 'C718e', 'C718mu', 'C718tau',
                              'C723u', 'C723d', 'C723s', 'C723c', 'C723b', 'C723e', 'C723mu', 'C723tau', 
                              'C725',
 
@@ -2327,10 +2264,6 @@ class WC_5flavor(object):
                                  'C716u', 'C716d', 'C716s', 'C716c', 'C716b', 'C716e', 'C716mu', 'C716tau',
                                  'C717u', 'C717d', 'C717s', 'C717c', 'C717b', 'C717e', 'C717mu', 'C717tau', 
                                  'C718u', 'C718d', 'C718s', 'C718c', 'C718b', 'C718e', 'C718mu', 'C718tau',
-                                 'C719u', 'C719d', 'C719s', 'C719c', 'C719b', 'C719e', 'C719mu', 'C719tau', 
-                                 'C720u', 'C720d', 'C720s', 'C720c', 'C720b', 'C720e', 'C720mu', 'C720tau', 
-                                 'C721u', 'C721d', 'C721s', 'C721c', 'C721b', 'C721e', 'C721mu', 'C721tau', 
-                                 'C722u', 'C722d', 'C722s', 'C722c', 'C722b', 'C722e', 'C722mu', 'C722tau',
                                  'C723u', 'C723d', 'C723s', 'C723c', 'C723b', 'C723e', 'C723mu', 'C723tau', 
                                  'C725']
 
@@ -2353,10 +2286,6 @@ class WC_5flavor(object):
                                     'C716u', 'C716d', 'C716s', 'C716c', 'C716e', 'C716mu', 'C716tau',
                                     'C717u', 'C717d', 'C717s', 'C717c', 'C717e', 'C717mu', 'C717tau', 
                                     'C718u', 'C718d', 'C718s', 'C718c', 'C718e', 'C718mu', 'C718tau',
-                                    'C719u', 'C719d', 'C719s', 'C719c', 'C719e', 'C719mu', 'C719tau', 
-                                    'C720u', 'C720d', 'C720s', 'C720c', 'C720e', 'C720mu', 'C720tau', 
-                                    'C721u', 'C721d', 'C721s', 'C721c', 'C721e', 'C721mu', 'C721tau', 
-                                    'C722u', 'C722d', 'C722s', 'C722c', 'C722e', 'C722mu', 'C722tau',
                                     'C723u', 'C723d', 'C723s', 'C723c', 'C723e', 'C723mu', 'C723tau', 
                                     'C725']
 
@@ -2369,10 +2298,6 @@ class WC_5flavor(object):
                                  'C77u', 'C77d', 'C77s', 'C77c', 'C77b', 'C77e', 'C77mu', 'C77tau', 
                                  'C78u', 'C78d', 'C78s', 'C78c', 'C78b', 'C78e', 'C78mu', 'C78tau',
                                  'C711', 'C712', 'C713', 'C714',
-                                 'C715u', 'C715d', 'C715s', 'C715c', 'C715b', 'C715e', 'C715mu', 'C715tau', 
-                                 'C716u', 'C716d', 'C716s', 'C716c', 'C716b', 'C716e', 'C716mu', 'C716tau',
-                                 'C717u', 'C717d', 'C717s', 'C717c', 'C717b', 'C717e', 'C717mu', 'C717tau', 
-                                 'C718u', 'C718d', 'C718s', 'C718c', 'C718b', 'C718e', 'C718mu', 'C718tau',
                                  'C723u', 'C723d', 'C723s', 'C723c', 'C723b', 'C723e', 'C723mu', 'C723tau', 
                                  'C725']
 
@@ -2380,7 +2305,7 @@ class WC_5flavor(object):
 
             # The list of indices to be deleted from the QCD/QED ADM because of less operators
             del_ind_list = [i for i in range(0,10)] + [i for i in range(18,26)]\
-                           + [i for i in range(70,86)] + [i for i in range(122,154)]
+                           + [i for i in range(70,86)] + [i for i in range(90,122)]
             # The list of indices to be deleted from the dim.8 ADM because of less operators
             del_ind_list_dim_8 = np.r_[np.s_[0:3], np.s_[6:9]]
             # The list of indices to be deleted from the ADT because of less operators (dim.6 part)
@@ -2395,10 +2320,6 @@ class WC_5flavor(object):
                                     'C77u', 'C77d', 'C77s', 'C77c', 'C77e', 'C77mu', 'C77tau', 
                                     'C78u', 'C78d', 'C78s', 'C78c', 'C78e', 'C78mu', 'C78tau',
                                     'C711', 'C712', 'C713', 'C714',
-                                    'C715u', 'C715d', 'C715s', 'C715c', 'C715e', 'C715mu', 'C715tau', 
-                                    'C716u', 'C716d', 'C716s', 'C716c', 'C716e', 'C716mu', 'C716tau',
-                                    'C717u', 'C717d', 'C717s', 'C717c', 'C717e', 'C717mu', 'C717tau', 
-                                    'C718u', 'C718d', 'C718s', 'C718c', 'C718e', 'C718mu', 'C718tau',
                                     'C723u', 'C723d', 'C723s', 'C723c', 'C723e', 'C723mu', 'C723tau', 
                                     'C725']
 
@@ -2417,7 +2338,7 @@ class WC_5flavor(object):
             # The list of indices to be deleted from the QCD/QED ADM because of less operators
             del_ind_list = [0,1] + [i for i in range(10,18)] + [i for i in range(26,34)]\
                            + [35] + [37] + [i for i in range(46,54)]\
-                           + [i for i in range(62,86)] + [87] + [89] + [i for i in range(90,154)]
+                           + [i for i in range(62,86)] + [87] + [89] + [i for i in range(90,122)]
             # The list of indices to be deleted from the dim.8 ADM because of less operators
             del_ind_list_dim_8 = np.r_[np.s_[0:3], np.s_[6:9]]
             # The list of indices to be deleted from the ADT because of less operators (dim.6 part)
@@ -2446,7 +2367,7 @@ class WC_5flavor(object):
             # The list of indices to be deleted from the QCD/QED ADM because of less operators
             del_ind_list = [i for i in range(0,34)] + [35] + [37] + [i for i in range(46,54)]\
                            + [i for i in range(62,86)]\
-                           + [87] + [89] + [i for i in range(90,154)]
+                           + [87] + [89] + [i for i in range(90,122)]
 
             # The 4-flavor list for matching only
             self.wc_name_list_4f = ['C65', 'C66',
@@ -3620,6 +3541,13 @@ class WilCo_EW(object):
                              'C78u', 'C78d', 'C78s', 'C68c', 'C68b', 'C78e', 'C78mu', 'C78tau',
                              'C79u', 'C79d', 'C79s', 'C69c', 'C69b', 'C79e', 'C79mu', 'C79tau', 
                              'C710u', 'C710d', 'C710s', 'C610c', 'C610b', 'C710e', 'C710mu', 'C710tau'
+                             'C711', 'C712', 'C713', 'C714',
+                             'C715u', 'C715d', 'C715s', 'C715e', 'C715mu', 'C715tau',
+                             'C716u', 'C716d', 'C716s', 'C716e', 'C716mu', 'C716tau',
+                             'C717u', 'C717d', 'C717s', 'C717e', 'C717mu', 'C717tau',
+                             'C718u', 'C718d', 'C718s', 'C718e', 'C718mu', 'C718tau',
+                             'C723u', 'C723d', 'C723s', 'C723e', 'C723mu', 'C723tau',
+                             'C725'
         """
         if RUN_EW is None:
             RUN_EW = True
@@ -4262,42 +4190,6 @@ class WilCo_EW(object):
         coeff_dict_5f['C718e'] = 0
         coeff_dict_5f['C718mu'] = 0
         coeff_dict_5f['C718tau'] = 0
-
-        coeff_dict_5f['C719u'] = 0
-        coeff_dict_5f['C719d'] = 0
-        coeff_dict_5f['C719s'] = 0
-        coeff_dict_5f['C719c'] = 0
-        coeff_dict_5f['C719b'] = 0
-        coeff_dict_5f['C719e'] = 0
-        coeff_dict_5f['C719mu'] = 0
-        coeff_dict_5f['C719tau'] = 0
-
-        coeff_dict_5f['C720u'] = 0
-        coeff_dict_5f['C720d'] = 0
-        coeff_dict_5f['C720s'] = 0
-        coeff_dict_5f['C720c'] = 0
-        coeff_dict_5f['C720b'] = 0
-        coeff_dict_5f['C720e'] = 0
-        coeff_dict_5f['C720mu'] = 0
-        coeff_dict_5f['C720tau'] = 0
-
-        coeff_dict_5f['C721u'] = 0
-        coeff_dict_5f['C721d'] = 0
-        coeff_dict_5f['C721s'] = 0
-        coeff_dict_5f['C721c'] = 0
-        coeff_dict_5f['C721b'] = 0
-        coeff_dict_5f['C721e'] = 0
-        coeff_dict_5f['C721mu'] = 0
-        coeff_dict_5f['C721tau'] = 0
-
-        coeff_dict_5f['C722u'] = 0
-        coeff_dict_5f['C722d'] = 0
-        coeff_dict_5f['C722s'] = 0
-        coeff_dict_5f['C722c'] = 0
-        coeff_dict_5f['C722b'] = 0
-        coeff_dict_5f['C722e'] = 0
-        coeff_dict_5f['C722mu'] = 0
-        coeff_dict_5f['C722tau'] = 0
 
         coeff_dict_5f['C725'] = 0