+++ ChangeLog.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [2.2.2] - 2023-07-10
+### Fixed
+- Key error in wilson_coefficients.py (`cNR13n` was not defined)
+- Removed spurious NLO terms in `cNR` coefficients
+- Deleted spurious dimension-seven Wilson coefficients. The basis now corresponds to [arxix [v3] of this paper](https://arxiv.org/abs/1710.10218).
+
+### Changed
+- `example.py` includes a few more examples for RG running below the weak scale
+
 ## [2.2.1] - 2023-04-21
 ### Fixed
 - Inconsistent implementation of double weak insertions for lepton operators